Redesigned redesigned how we create sub users config data
